Don't write python bytecode while testing
This should reduce the number of times you need to do things like git
clean -x -i to kill pyc files in your local directory to get tests to
pass.
It doesn't appear to address pyc files in the .tox directories, and
attempts to address that seem to go in circles. Ideas welcomed.

Change-Id: Id01624a8d21c02679c7ba659232cf27323fafdc0
Related-Bug: #1368661