Allow alternate OpenStack themes
By default oslosphinx forces projects that use it to use the theme
contained in the package. The documentation team wants to provide a
separate theme for some parts of the docs.openstack.org site. They are
going to distribute it in their own package, but do still want to be
able to use other features of oslosphinx. This change allows any name
that starts "openstack" to be used, while retaining the default behavior
of inserting the "openstack" theme if a project is using a different
default or has not set a theme.

Change-Id: I7a2c5a5c3805685b8fa0baaa47393d5fec312f30