Remove google analytics
This patch removes the inclusion of google-analytics, which has nothing
to do in a free software project like OpenStack. In no way, we should
be including such a spy tool on all of our documentations. It is by the
way considered as a privacy breach and a very bad practice by
distributions like Debian, so at least there, the OpenStack theme has
to be patched anyway.

Change-Id: I4867292dcfb1a4d13822076208b349130e5c93b0