Fix python3 compatibility in _html_page_context
Trying to build a project documentation (e.g. os-testr) using
sphinx-build on a python3 venv failed with the following error
message:

File "...oslosphinx/__init__.py", line 57, in _html_page_context
    other_versions = [t for t in raw_version_list.split('\n')
TypeError: 'str' does not support the buffer interface

This is because raw_version_list is of type str. The fix makes it
compatible with python2 and python3.

Change-Id: I934260435d1225418eb2fe78dc096ba62124a22f