Allow "Other Versions" section to be configurable
Adds a new option that allows the "Other Versions" section to be removed
if a project isn't interested in showing links to older versions. The
default value is False which means that it will be hidden.

Even though the setting is not backward compatible there are enough
broken documentation links in the OpenStack projects that this should be
opt-in behavior.

Change-Id: Ic4b47b19aa9db38557c37aed7722174b7dbdf4a1
Closes-Bug: #1599260