RTD is not rendering toctree bullets
The Read the Docs sphinx build is not rendering the bullets of a
toctree. This results in a flat list of items where are indented
list should be show.

The root cause is, that the conf.py file gets modfied by RTD which
set sets the 'html_theme' option of the 'html_context' dict to
the rtd theme. The oslosphinx extension is not overwriting this.

This patch ensures that 'html_theme' in the 'html_context' dict
is set to 'opentack' as well.

Change-Id: I18b0af3cc75e560cc87e49b38e6f867a10e4a69a
Closes-Bug: #1664976