Amended dev documentation side navigation
Amended developer documentation side navigation to add an option for
clearly returning to the project developer documentation home page.
Previously it was possible to return to the home page by clicking
'table of contents' however this is not obvious and makes it hard
to get around the complete project documentation once you have left
the homepage.

This patch removes the link from 'table of contents' which is
misleading, and a link is added beneath this header to say 'Return
to project home page'.

Change-Id: I7cf96acead62f00224ca9217a80760b4a65d42d6
Closes-bug: #1666323